Changed sizeof() to array size.
We noticed that the file names were truncated at eight characters.  The
root cause was using sizeof(name) instead of strlen(name).  The sizeof()
operator was returning the size of the pointer (8 bytes on a 64-bit
binary) instead of the length of the string of the overall size of the
array.