Refresh stats page every 30 seconds
I thought this might be useful for everyone. 

First tries to use the meta tag option to refresh the page made it unusable. So I had to find a different option. This javascript function should work in most browsers. 
It would probably also be possible to add a selector to choose between no refresh and a different timing of 5/10/30/60 seconds or similar. I could take a look at this next if you like the idea.