Add "record_align_interval" option
If interval recording enabled, the option makes the module to
recreate the dump file precisely at the interval boundaries, and
the first interval can be trancated by the boundary.

E.g. if you start recording at 10:10:15 and the interval is 30
seconds, if the option is "off", the files will be created at
10:10:45, 10:11:15 and so on. If the option is enabled, the times
are 10:10:30, 10:11:00, 10:11:30 etc.